FASE 1: Eliminar sistema de scrolling completamente
Eliminado el sistema de scrolling de niri para preparar la conversión
a un gestor de ventanas estilo i3. Esto incluye:

- Comentados todos los input handlers de scroll (wheel y touchpad)
- Eliminados bindings de scroll en configuración
- Reemplazado scrolling.rs (5000+ líneas) con stub mínimo (459 líneas)
- Corregidos todos los errores de compilación (134 → 0)
- Backup del scrolling original guardado en scrolling.rs.BACKUP

El stub temporal permite que el código compile mientras se implementa
el sistema de contenedores i3 en FASE 2.

Estado: ✅ 0 errores de compilación, 7 warnings esperados (código muerto)

+++ src/layout/floating.rs
@@ -530,7 +530,7 @@ impl<W: LayoutElement> FloatingSpace<W> {
         // Store the floating position.
         tile.floating_pos = Some(data.pos);
 
-        let width = ColumnWidth::Fixed(tile.tile_expected_or_current_size().w);
+        let width = ColumnWidth::Fixed(tile.tile_expected_or_current_size().w as i32);
         RemovedTile {
             tile,
             width,
